The code for comparing the Hashes
a for loop over the keys of one of the hashes. As the names of
the keys is the same for both, I can use that to access the
probe name for the reverse complement and the forward contig.
From there I can compare the two and push the name onto the
approprate hash.