Testing hash of arrays formation
Turns out I didn't really need all of that control structure
from the previous commit because perl is clever and will
create a new key if one doesn't exist and make a new array to
keep my probes in.

Also added some Dumper code so that I could check that the hash
of arrays contained what I expected as I expect it to.